Refactor
Changes:
- provide QiitaService by Dagger
- Extract User class from Feed class
- Remove unnecessary code
- Make fragment care only UI thread
