+++ app/build.gradle
@@ -36,7 +36,6 @@ android {
 }
 
 dependencies {
-    implementation fileTree(dir: 'libs', include: ['*.jar'])
     implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
 
     // unit tests - https://developer.android.com/training/testing/unit-testing/local-unit-tests
