+++ app/src/main/java/com/example/alpacamusclemaintenance/viewmodel/FeedViewModel.kt
@@ -4,6 +4,7 @@ import androidx.lifecycle.ViewModel
 import com.example.alpacamusclemaintenance.repository.FeedRepository
 import com.example.alpacamusclemaintenance.vo.Feed
 import io.reactivex.Single
+import io.reactivex.schedulers.Schedulers
 import javax.inject.Inject
 
 class FeedViewModel @Inject constructor(
@@ -11,5 +12,7 @@ class FeedViewModel @Inject constructor(
 ) : ViewModel() {
 
   val feedObservable: Single<List<Feed>>
-    get() = repository.getFeedsByTag("筋トレ")
+    get() = repository
+      .getFeedsByTag("筋トレ")
+      .subscribeOn(Schedulers.io())
 }
