+++ app/src/main/java/com/example/alpacamusclemaintenance/vo/Feed.kt
@@ -3,18 +3,15 @@ package com.example.alpacamusclemaintenance.vo
 import java.util.*
 
 data class Feed(
-  var user: User,
-  var body: String,
-  var createdAt: Date,
-  var likesCount: Int,
-  var title: String,
-  var url: String
-) {
+  val user: User,
+  val body: String,
+  val createdAt: Date,
+  val likesCount: Int,
+  val title: String,
+  val url: String
+)
 
-  data class User(
-    var id: String,
-    var profileImageUrl: String
-  )
-
-}
-
+data class User(
+  val id: String,
+  val profileImageUrl: String
+)
