+++ app/src/main/java/com/example/alpacamusclemaintenance/vo/Home.kt
@@ -1,7 +1,6 @@
 package com.example.alpacamusclemaintenance.vo
 
 data class Home(
-  var wordOfWisdom: String,
-  var author: String
+  val wordOfWisdom: String,
+  val author: String
 )
-
